+++ src/render/render_cub3d.c
@@ -6,7 +6,7 @@
 /*   By: joseoliv <joseoliv@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/01 14:18:14 by joseoliv          #+#    #+#             */
-/*   Updated: 2025/03/22 02:27:07 by joseoliv         ###   ########.fr       */
+/*   Updated: 2025/03/22 05:50:31 by joseoliv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 int	render_cub3d(void *param)
 {
 	t_game	*game;
-	float	fraction;
-	float	start_x;
+	double	fraction;
+	double	start_x;
 
 	game = (t_game *)param;
 	move_player(game);
