+++ src/utils/raycaster.utils.c
@@ -3,33 +3,33 @@
 /*                                                        :::      ::::::::   */
 /*   raycaster.utils.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cereais <cereais@student.42.fr>            +#+  +:+       +#+        */
+/*   By: joseoliv <joseoliv@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/14 02:56:34 by joseoliv          #+#    #+#             */
-/*   Updated: 2025/03/21 07:50:28 by cereais          ###   ########.fr       */
+/*   Updated: 2025/03/22 05:50:32 by joseoliv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-void	set_ray_parameters(t_game *game, float *fraction, float *start_x)
+void	set_ray_parameters(t_game *game, double *fraction, double *start_x)
 {
 	*fraction = PI / 4 / WIDTH;
 	*start_x = game->player.angle - (PI / 4) / 2;
 }
 
-void	calculate_ray_direction(float start_x, t_ray *ray)
+void	calculate_ray_direction(double start_x, t_ray *ray)
 {
 	ray->dir_x = cosf(start_x);
 	ray->dir_y = sinf(start_x);
 }
 
-double	degree_to_radians(float a)
+double	degree_to_radians(double a)
 {
 	return (a * PI / 180.0);
 }
 
-bool	touch_wall(float px, float py, t_game *game)
+bool	touch_wall(double px, double py, t_game *game)
 {
 	if (px < 0 || px >= WIDTH || py < 0 || py >= HEIGHT)
 		return (true);
